Add `RepositoryDetails`, and `ResolutionAwareRepository#getDetails()` and `ModuleComponentRepository#getDetails`
This allows to use whatever we want in `RepositoryDetails`, such as a different ID, really unique across all repositories of a given configuration.
This requirement is needed to safely link a `ResolvedComponentResult` to _the_ repository that was used as source.

The existing ID used by Gradle during resolution is on purpose not that fine-grained (to allow e.g. an artifact that was downloaded by a repo named 'foo' to be kept, and not redownloaded if the repo is renamed 'bar')